Move get_signature_recovery_value() to the test where it is used
